Metrics page basics
